args: "explode" arguments containing non-basic runes
... so that it's clearer to the user which runes are actually making up
the argument, if it's non-"basic" (i.e. 0x20-0x7e plus a few control
characters).

Also, provide "good" names for "control characters" that
`runenames.Name(r)` insists on "just" calling `<control>`.

Example:

    $ go run . args foo ààà
    2 ARGV:
    1	foo
    2	ààà
      - 224  0x00e0 LATIN SMALL LETTER A WITH GRAVE (graphic, letter, lower, printable)
      - 97   0x0061 LATIN SMALL LETTER A (graphic, letter, lower, printable)
      - 768  0x0300 COMBINING GRAVE ACCENT (graphic, mark, printable)
      - 224  0x00e0 LATIN SMALL LETTER A WITH GRAVE (graphic, letter, lower, printable)

In the above example, the first argument _isn't_ expanded, as it's
comprised only by "basic" runes. The second isn't, so it's exploded -
revealing that the two seemingly "same" letters were in fact made up of
different runes (the first, a rune on its own; the second, a combination
of a standard letter and a combining character).